Load initial setup
This update has all of the rough form starting to show. The getters and
setters are all in place for the class, and the main framework for the
XAML screen, and the mainpage program have been run in. GPIO
initialization has started. The output pins are defined. Need to work on
the input pins.